Switch to Vertico, Orderless, Marginalia, Consult, and Embark.
(And also, move back to isearch!)

Note that instead of using `persp-switch-to-buffer*' for buffer switching, I
added the perspective-filtered buffer list as a virtual buffer source to
Consult. This puts the list of buffers in the current perspective at the top of
the list of the `consult-buffer' buffer switcher. Technically buffers that are
unrelated to the current perspective appear, however they are clearly
demarcated, and I like the flexibility of easily being able to switch to any
buffer, regardless of perspective. And, as an extra bonus, other Consult
virtual buffer sources show up in all contexts, which I find really handy.

This transition replaces Ivy, Swiper, and Counsel. (This commit also removes
some old code related to ido, flx, smex, etc. that I haven't been using.)